Fix mocking of rIC in tests
**what is the change?:**
In our tests we were relying on the fact that the 'native' rIC and rAF
were being used. These were either manually polyfilled, or we relied on
jest/node polyfilling them.

We're making our own scheduler now, and in general it should be mocked
in tests of React, and for users writing tests.

I tried to pull the mocking logic into a manual mock under
`packages/react-scheduler/src/__mocks__/ReactScheduler.js` but no matter
what I did it wasn't being used in the tests. So for now inlining the
logic because that works and we want to keep moving.

**why make this change?:**
Get tests passing ✨

**test plan:**
run the tests

**issue:**
internal task T28128480